renamed and adapted forecasting models tests
